+++ src/components/dropdown.tsx
@@ -1,21 +1,21 @@
 import { useState } from "react";
-import Icons from "./Icons";
+import {Icon} from "./Icons";
 type Dropdownprops = {
-    icon?: string;
+    icon?: Icon;
     onClick?: () => void;
     options: Array<{ text: string; onClick: () => void }>;
 }
-export default function Dropdown({ options, icon, onClick }: Dropdownprops) {
+export default function Dropdown(props: Dropdownprops) {
     const [isOpen, setIsOpen] = useState(false);
     function toggleDropdown() {
         setIsOpen(!isOpen);
     }
     return (
         <div className="relative">
-            {icon && Icons[icon]({onClick})}
+            {(props.icon) ? <props.icon/> : ''}
             <div
                 className={`dropdown-menu absolute bg-white shadow-md w-auto rounded mt-1 z-10 front-layer ${isOpen ? "" : "hidden"}`}>
-                {options.map((option, index) => (
+                {props.options.map((option, index) => (
                     <a key={option.text} href="#" className="block px-2 py-2 w-28 text-gray-800 hover:bg-gray-100" onClick={option.onClick} >
                         {option.text}
                     </a>))}
